whatpm/t/xml/doctypes-1.dat

#data
<!DOCTYPE abcd><abcd/>
#errors
#document
| <!DOCTYPE abcd>
| <{} abcd>

#data
<!DOCTYPE Abcd><Abcd/>
#errors
#document
| <!DOCTYPE Abcd>
| <{} Abcd>

#data
<!DOCTYPE ABCD><ABCD/>
#errors
#document
| <!DOCTYPE ABCD>
| <{} ABCD>

#data
<!DOCTYPE HtML>
<HtML/>
#errors
#document
| <!DOCTYPE HtML>
| <HtML>

#data
<!DOCTYPE html>
<HTML/>
#errors
#document
| <!DOCTYPE html>
| <HTML>
